refactored showTab function, fixed minifier
At some point, I changed the minifier to use : instead of = to separate
keys and values, but hadn't updated the deployment script, so the full,
unminified version of the site was getting deployed. suuuuuuuck.

Also, refactored the showTab function to be a shorter function call in
the general use case.